refactor(#664): refactor hanami for the new torii
Refactoring of hanami by pulling out token-validation
stuff. Replaced the token-validation for now by a
simple extraction.
