+++ src/binaries/hanami/src/api/http_endpoints/checkpoint/delete_checkpoint_v1_0.rs
@@ -19,10 +19,10 @@ use apistos::actix::NoContent;
 use apistos::api_operation;
 use uuid::Uuid;
 
-use crate::api::errors::ErrorResponse;
-use crate::api::user_context::UserContext;
 use crate::database::checkpoint_table;
 
+use ainari_api::errors::ErrorResponse;
+use ainari_api::user_context::UserContext;
 use ainari_common::enums;
 
 #[api_operation(
