+++ src/binaries/hanami/src/api/http_endpoints/cluster/create_cluster_v1_0.rs
@@ -18,12 +18,12 @@ use apistos::api_operation;
 use uuid::Uuid;
 use validator::Validate;
 
-use crate::api::errors::ErrorResponse;
-use crate::api::user_context::UserContext;
 use crate::core::cluster_handler::CLUSTER_HANDLER;
 use crate::database::cluster_table;
 
+use ainari_api::errors::ErrorResponse;
 use ainari_api::structs::cluster_structs::{ClusterCreateReq, ClusterResp};
+use ainari_api::user_context::UserContext;
 use ainari_common::error::AinariError;
 
 #[api_operation(
