+++ src/binaries/hanami/src/api/http_endpoints/cluster/delete_cluster_v1_0.rs
@@ -17,11 +17,11 @@ use apistos::actix::NoContent;
 use apistos::api_operation;
 use uuid::Uuid;
 
-use crate::api::errors::ErrorResponse;
-use crate::api::user_context::UserContext;
 use crate::core::cluster_handler;
 use crate::database::cluster_table;
 
+use ainari_api::errors::ErrorResponse;
+use ainari_api::user_context::UserContext;
 use ainari_common::enums;
 use ainari_common::error::AinariError;
