+++ src/binaries/hanami/src/api/http_endpoints/cluster/get_cluster_v1_0.rs
@@ -17,11 +17,11 @@ use actix_web::web::Path;
 use apistos::api_operation;
 use uuid::Uuid;
 
-use crate::api::errors::ErrorResponse;
-use crate::api::user_context::UserContext;
 use crate::database::cluster_table;
 
+use ainari_api::errors::ErrorResponse;
 use ainari_api::structs::cluster_structs::ClusterResp;
+use ainari_api::user_context::UserContext;
 use ainari_common::enums;
 
 #[api_operation(
