+++ src/binaries/hanami/src/api/http_endpoints/cluster/task/abort_task_v1_0.rs
@@ -18,12 +18,12 @@ use apistos::api_operation;
 use std::str::FromStr;
 use uuid::Uuid;
 
-use crate::api::errors::ErrorResponse;
-use crate::api::user_context::UserContext;
 use crate::database::cluster_table;
 use crate::database::task_table;
 
+use ainari_api::errors::ErrorResponse;
 use ainari_api::structs::task_structs::{TaskResp, TaskState, TaskType};
+use ainari_api::user_context::UserContext;
 use ainari_common::enums;
 
 #[api_operation(
