+++ src/binaries/hanami/src/api/http_endpoints/cluster/train_cluster_v1_0.rs
@@ -19,12 +19,12 @@ use apistos::api_operation;
 use uuid::Uuid;
 use validator::Validate;
 
-use crate::api::errors::ErrorResponse;
-use crate::api::user_context::UserContext;
 use crate::core::cluster_handler;
 use crate::database::cluster_table;
 
+use ainari_api::errors::ErrorResponse;
 use ainari_api::structs::cluster_structs::ClusterTrainReq;
+use ainari_api::user_context::UserContext;
 use ainari_common::enums;
 use ainari_common::error::AinariError;
