+++ src/binaries/hanami/src/api/http_endpoints/dataset/list_dataset_v1_0.rs
@@ -16,11 +16,11 @@ use actix_web::web::Json;
 use apistos::api_operation;
 use uuid::Uuid;
 
-use crate::api::errors::ErrorResponse;
-use crate::api::user_context::UserContext;
 use crate::database::dataset_table;
 
+use ainari_api::errors::ErrorResponse;
 use ainari_api::structs::dataset_structs::{DatasetBasicResp, DatasetListResp};
+use ainari_api::user_context::UserContext;
 
 #[api_operation(
     tag = "dataset",
