+++ src/binaries/hanami/src/api/mod.rs
@@ -12,9 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-pub mod errors;
 pub mod http_endpoints;
 pub mod http_server;
 pub mod middleware;
-pub mod token_handling;
-pub mod user_context;
