+++ src/libs/rust/ainari_api/Cargo.toml
@@ -18,11 +18,11 @@ apistos = { version = "0.6", features = ["extras"] }
 validator = { version = "0.20", features = ["derive"] }
 # json
 serde = { version = "1", features = ["derive"] }
+serde_json = "1"
 # common stuff
 uuid = { version = "1.18.0", features = ["v4"] }
 futures = "0.3"
+base64 = "0.22" 
 # singleton (like used for the config-variable)
 once_cell = "1.18"
-# jwt-token
-jsonwebtoken = "9"
